Refactor sqlalchemy repository tests
