+++ tests/contrib/sqlalchemy/repository/conftest.py
@@ -2,10 +2,14 @@ from __future__ import annotations
 
 import json
 from datetime import datetime
-from typing import TYPE_CHECKING, Any
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, AsyncGenerator, Generator
 from uuid import UUID
 
 import pytest
+from sqlalchemy import URL, Engine, NullPool, create_engine
+from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, async_sessionmaker, create_async_engine
+from sqlalchemy.orm import Session, sessionmaker
 
 if TYPE_CHECKING:
     from pytest import MonkeyPatch
@@ -175,3 +179,232 @@ def fx_raw_rules_bigint() -> list[dict[str, Any]]:
             "updated": "2023-02-01T00:00:00",
         },
     ]
+
+
+@pytest.fixture()
+def duckdb_engine(tmp_path: Path) -> Generator[Engine, None, None]:
+    """SQLite engine for end-to-end testing.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    engine = create_engine(f"duckdb:///{tmp_path}/test.duck.db", poolclass=NullPool)
+    try:
+        yield engine
+    finally:
+        engine.dispose()
+
+
+@pytest.fixture()
+def oracle_engine(docker_ip: str, oracle_service) -> Engine:
+    """Postgresql instance for end-to-end testing.
+
+    Args:
+        docker_ip: IP address for TCP connection to Docker containers.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    return create_engine(
+        "oracle+oracledb://:@",
+        thick_mode=False,
+        connect_args={
+            "user": "app",
+            "password": "super-secret",
+            "host": docker_ip,
+            "port": 1512,
+            "service_name": "xepdb1",
+            "encoding": "UTF-8",
+            "nencoding": "UTF-8",
+        },
+        poolclass=NullPool,
+    )
+
+
+@pytest.fixture()
+def psycopg_engine(docker_ip: str, postgres_service) -> Engine:
+    """Postgresql instance for end-to-end testing.
+
+    Args:
+        docker_ip: IP address for TCP connection to Docker containers.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    return create_engine(
+        URL(
+            drivername="postgresql+psycopg",
+            username="postgres",
+            password="super-secret",
+            host=docker_ip,
+            port=5423,
+            database="postgres",
+            query={},  # type:ignore[arg-type]
+        ),
+        poolclass=NullPool,
+    )
+
+
+@pytest.fixture()
+def sqlite_engine(tmp_path: Path) -> Generator[Engine, None, None]:
+    """SQLite engine for end-to-end testing.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    engine = create_engine(f"sqlite:///{tmp_path}/test.db", poolclass=NullPool)
+    try:
+        yield engine
+    finally:
+        engine.dispose()
+
+
+@pytest.fixture()
+def spanner_engine(docker_ip: str, spanner_service, monkeypatch: MonkeyPatch) -> Engine:
+    """Postgresql instance for end-to-end testing.
+
+    Args:
+        docker_ip: IP address for TCP connection to Docker containers.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    monkeypatch.setenv("SPANNER_EMULATOR_HOST", "localhost:9010")
+    monkeypatch.setenv("GOOGLE_CLOUD_PROJECT", "emulator-test-project")
+
+    return create_engine(
+        "spanner+spanner:///projects/emulator-test-project/instances/test-instance/databases/test-database"
+    )
+
+
+@pytest.fixture(
+    params=[
+        pytest.param("duckdb_engine", marks=pytest.mark.sqlalchemy_duckdb),
+        pytest.param("oracle_engine", marks=pytest.mark.sqlalchemy_oracledb),
+        pytest.param("psycopg_engine", marks=pytest.mark.sqlalchemy_psycopg_sync),
+        pytest.param("sqlite_engine", marks=pytest.mark.sqlalchemy_sqlite),
+    ]
+)
+def engine(request) -> Engine:
+    return request.getfixturevalue(request.param)
+
+
+@pytest.fixture()
+def session(engine: Engine) -> Generator[Session, None, None]:
+    session = sessionmaker(bind=engine)()
+    try:
+        yield session
+    finally:
+        session.rollback()
+        session.close()
+
+
+@pytest.fixture()
+async def engine_aiosqlite(tmp_path: Path) -> AsyncGenerator[AsyncEngine, None]:
+    """SQLite engine for end-to-end testing.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    engine = create_async_engine(f"sqlite+aiosqlite:///{tmp_path}/test.db", poolclass=NullPool)
+    try:
+        yield engine
+    finally:
+        await engine.dispose()
+
+
+@pytest.fixture()
+async def engine_asyncmy(docker_ip: str, mysql_service) -> AsyncEngine:
+    """Postgresql instance for end-to-end testing.
+
+    Args:
+        docker_ip: IP address for TCP connection to Docker containers.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    return create_async_engine(
+        URL(
+            drivername="mysql+asyncmy",
+            username="app",
+            password="super-secret",
+            host=docker_ip,
+            port=3360,
+            database="db",
+            query={},  # type:ignore[arg-type]
+        ),
+        echo=True,
+        poolclass=NullPool,
+    )
+
+
+@pytest.fixture()
+async def engine_asyncpg(docker_ip: str, postgres_service) -> AsyncEngine:
+    """Postgresql instance for end-to-end testing.
+
+    Args:
+        docker_ip: IP address for TCP connection to Docker containers.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    return create_async_engine(
+        URL(
+            drivername="postgresql+asyncpg",
+            username="postgres",
+            password="super-secret",
+            host=docker_ip,
+            port=5423,
+            database="postgres",
+            query={},  # type:ignore[arg-type]
+        ),
+        echo=True,
+        poolclass=NullPool,
+    )
+
+
+@pytest.fixture()
+async def engine_psycopg_async(docker_ip: str, postgres_service) -> AsyncEngine:
+    """Postgresql instance for end-to-end testing.
+
+    Args:
+        docker_ip: IP address for TCP connection to Docker containers.
+
+    Returns:
+        Async SQLAlchemy engine instance.
+    """
+    return create_async_engine(
+        URL(
+            drivername="postgresql+psycopg",
+            username="postgres",
+            password="super-secret",
+            host=docker_ip,
+            port=5423,
+            database="postgres",
+            query={},  # type:ignore[arg-type]
+        ),
+        echo=True,
+        poolclass=NullPool,
+    )
+
+
+@pytest.fixture(
+    params=[
+        pytest.param("engine_aiosqlite", marks=pytest.mark.sqlalchemy_aiosqlite),
+        pytest.param("engine_asyncmy", marks=pytest.mark.sqlalchemy_asyncmy),
+        pytest.param("engine_asyncpg", marks=pytest.mark.sqlalchemy_asyncpg),
+        pytest.param("engine_psycopg_async", marks=pytest.mark.sqlalchemy_psycopg_async),
+    ]
+)
+def async_engine(request) -> AsyncEngine:
+    return request.getfixturevalue(request.param)
+
+
+@pytest.fixture()
+async def async_session(async_engine: AsyncEngine) -> AsyncGenerator[AsyncSession, None]:
+    session = async_sessionmaker(bind=async_engine)()
+    try:
+        yield session
+    finally:
+        await session.rollback()
+        await session.close()
