+++ tests/contrib/sqlalchemy/repository/test_uuid_async.py
@@ -1,6 +1,7 @@
 """Unit tests for the SQLAlchemy Repository implementation."""
 from __future__ import annotations
 
+import sys
 from datetime import datetime
 from typing import Any
 from uuid import UUID, uuid4
@@ -9,6 +10,7 @@ import pytest
 from sqlalchemy import insert
 from sqlalchemy.ext.asyncio import (
     AsyncEngine,
+    AsyncSession,
 )
 
 from litestar.contrib.repository.exceptions import RepositoryError
@@ -22,18 +24,19 @@ from tests.contrib.sqlalchemy.models_uuid import (
     UUIDRule,
 )
 
+pytestmark = [
+    pytest.mark.skipif(sys.platform != "linux", reason="docker not available on this platform"),
+    pytest.mark.sqlalchemy_integration,
+]
 
+
+@pytest.fixture(autouse=True)
 async def seed_db(
-    engine: AsyncEngine,
+    async_engine: AsyncEngine,
     raw_authors_uuid: list[dict[str, Any]],
     raw_books_uuid: list[dict[str, Any]],
     raw_rules_uuid: list[dict[str, Any]],
 ) -> None:
-    """Populate test database with sample data.
-
-    Args:
-        engine: The SQLAlchemy engine instance.
-    """
     # convert date/time strings to dt objects.
     for raw_author in raw_authors_uuid:
         raw_author["dob"] = datetime.strptime(raw_author["dob"], "%Y-%m-%d").date()
@@ -43,11 +46,26 @@ async def seed_db(
         raw_author["created"] = datetime.strptime(raw_author["created"], "%Y-%m-%dT%H:%M:%S")
         raw_author["updated"] = datetime.strptime(raw_author["updated"], "%Y-%m-%dT%H:%M:%S")
 
-    async with engine.begin() as conn:
+    async with async_engine.begin() as conn:
         await conn.run_sync(base.orm_registry.metadata.drop_all)
         await conn.run_sync(base.orm_registry.metadata.create_all)
         await conn.execute(insert(UUIDAuthor).values(raw_authors_uuid))
         await conn.execute(insert(UUIDRule).values(raw_rules_uuid))
+
+
+@pytest.fixture(name="author_repo")
+def fx_author_repo(async_session: AsyncSession) -> AuthorAsyncRepository:
+    return AuthorAsyncRepository(session=async_session)
+
+
+@pytest.fixture(name="book_repo")
+def fx_book_repo(async_session: AsyncSession) -> BookAsyncRepository:
+    return BookAsyncRepository(session=async_session)
+
+
+@pytest.fixture(name="rule_repo")
+def fx_rule_repo(async_session: AsyncSession) -> RuleAsyncRepository:
+    return RuleAsyncRepository(session=async_session)
 
 
 def test_filter_by_kwargs_with_incorrect_attribute_name(author_repo: AuthorAsyncRepository) -> None:
